service report updations

+++ database/migrations/2025_05_19_034423_add_reference_no_to_demo_requests_table.php
@@ -19,10 +19,16 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
-        Schema::table('demo_requests', function (Blueprint $table) {
-            $table->dropColumn('reference_no');
-        });
-    }
+   /**
+ * Reverse the migrations.
+ */
+        public function down(): void
+        {
+            if (Schema::hasColumn('demo_requests', 'reference_no')) {
+                Schema::table('demo_requests', function (Blueprint $table) {
+                    $table->dropColumn('reference_no');
+                });
+            }
+        }
+
 };
